pyproject.toml
[project]
name = "sywi-data-platform"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    # Minimal dependencies for dev CLI utility commands
    "dagster",
    "dagster-webserver",
    "dagster-postgres",
    "dagster-dg-cli",
    "click>=8,<8.2",
    "duckdb",
    "pyyaml>=6.0.3",
    "typer>=0.9.0",
    "python-dotenv",
]

[project.scripts]
dev = "dev:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
only-include = ["dev.py"]

[tool.uv]
package = true

[dependency-groups]
dev = []
